put text in teacher panel indicating whether fully locked
